Create README.md
People's health is deteriorating day by day, it could be the fact of less nutrition in our diets causing more health issues, or many new diseases are affecting human health, one of the significant factors growing age of the population which also comes with many illnesses, also medical expenses are getting too high, due to manpower and resources shortages, and many other possible factors, which comes with a rising demand of health insurance, to fulfill the health insurance demand and keep insurance companies profitable, it is necessary to predict the accurate medical expenses according to the different category of patents, based on which health insurance package cost can be decided and offered to the customers based on their category profiles. To forecast accurate medical expenses, we will use the OLS algorithm to reduce the forecast error and the sum of the squared error of a model.